Move endorse out of HasPrivilege trait + testing
Moves the `endorse` function out of the HasPrivilege trait, since
HasPrivilege does not necessarily require something like an "integrity"
component that allows "endorsing". Instead, HasPrivilege has a
`downgrade_to` method that downgrades the label to some target label
using privilege if possible. Endorse is moved to be a method of DCLabel,
and tested to be equivalent.

This commit also adds quickcheck property testing to DCLabels, which
helped test equivalence of endorse and downgrade_to. It really shouldn't
be in this commit, but it is.